// hw_21/patterns/utils/types.ts
export enum NAVIGATION_ITEMS {
  CATALOG = 'Каталог',
  NEWS = 'Новости',
  CAR_FLEA_MARKET = 'Автобарахолка',
  HOUSES_AND_FLATS = 'Дома и квартиры',
  SERVICES = 'Услуги',
  FLEA_MARKET = 'Барахолка',
  FORUM = 'Форум',
}

export enum QUERY_STRING {
  LAPTOPS = 'Ноутбуки',
  PHONES = 'Мобильные телефоны',
  JUICERS = 'Соковыжималки',
}

export enum SELECTOR_TYPES {
  XPATH = 'xpath',
  CLASSNAME = 'className',
}

export enum PAGES {
  HOME = 'Home',
  CATALOG = 'Catalog',
  FORUM = 'Forum',
}